Refactoring the tests to be more practical
