+++ gpu_ocean/tests/random_numbers.py
@@ -7,11 +7,12 @@ import xmlrunner
 # $ sudo easy_install unittest-xml-reporting
 
 #import testUtils
-from dataAssimilation.OceanStateNoise_test import OceanStateNoiseTest
+from stochastic.OceanStateNoise_test import OceanStateNoiseTest
+from stochastic.RandomNumbers_test import RandomNumbersTest
 
 def printSupportedTests():
     print ("Supported tests:")
-    print ("0: All, 1: OceanStateNoise")
+    print ("0: All, 1: RandomNumbers, 2: OceanStateNoise")
 
 
 if (len(sys.argv) < 2):
@@ -34,8 +35,10 @@ if (jenkins):
 # Define the tests that will be part of our test suite:
 test_classes_to_run = None
 if tests == 0:
-    test_classes_to_run = [OceanStateNoiseTest]
+    test_classes_to_run = [RandomNumbersTest, OceanStateNoiseTest]
 elif tests == 1:
+    test_classes_to_run = [RandomNumbersTest]
+elif tests == 2:
     test_classes_to_run = [OceanStateNoiseTest]
 else:
     print ("Error: " + str(tests) + " is not a supported test number...")
